Move FILE_GLOBS and discoverFiles out of runner.ts into discover.ts verbatim. Pure relocation, no behaviour change; runner.ts was at the 200-line limit, and discovery is a distinct responsibility.
